DEX-350 Releasing process
SBT:
* release task now doing all work to prepare a release;
* packageAll just packages artifacts;
* artifactExtensions are used to search artifacts;
* writeReleaseNotes writes release notes including artifacts hashes;
