+++ build.sbt
@@ -1,9 +1,13 @@
+import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Path}
 
 import CommonSettings.autoImport.network
+import Hashes.mk
+import org.apache.commons.codec.binary.Hex
 import sbt.Keys._
 import sbt._
 import sbt.internal.inc.ReflectUtilities
+import sbt.io.Using
 
 addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.0" cross CrossVersion.full)
 
@@ -117,13 +121,20 @@ genDocs := Def.taskDyn {
   (dex / Compile / runMain).toTask(s" com.wavesplatform.dex.MatcherTool ${configFile.getAbsolutePath} gen-docs ${releaseDirectory.value}")
 }.value
 
+lazy val release = taskKey[Unit]("Packages artifacts and writes a documentation to the releaseDirectory")
+release := Def
+  .sequential(
+    root / cleanAll,
+    Def.task { Files.createDirectories(releaseDirectory.value) },
+    Def.task { Command.process("packageAllForNetworks", state.value) },
+    genDocs,
+    writeReleaseNotes
+  )
+  .value
+
 lazy val packageAll = taskKey[Unit]("Package all artifacts")
 packageAll := Def
   .sequential(
-    root / cleanAll,
-    Def.task {
-      Files.createDirectories(releaseDirectory.value)
-    },
     Def.task {
       val artifacts = Seq(
         (dex / Universal / packageZipTarball).value,
@@ -131,17 +142,65 @@ packageAll := Def
       )
 
       val destDir = releaseDirectory.value
-      ReleaseHelpers.writeHashesFile(destDir.resolve("checksums.md").toFile, artifacts)
-
       artifacts
         .map(_.toPath)
         .foreach { source =>
           Files.move(source, destDir.resolve(source.getFileName))
         }
-    },
-    genDocs
+    }
   )
   .value
+
+lazy val artifactExtensions = settingKey[Seq[String]]("Used to search artifacts")
+artifactExtensions := List("deb", "tgz")
+
+lazy val writeReleaseNotes = taskKey[Unit](s"""1. Collects commits between last two tags
+     |2. Writes a draft of a release to the releaseDirectory
+     |3. Adds checksums for artifacts in the releaseDirectory""".stripMargin)
+writeReleaseNotes := {
+  val runner           = git.runner.value
+  val log              = streams.value.log
+  val cwd              = (root / baseDirectory).value
+  val destDir          = (root / releaseDirectory).value
+  val releaseNotesFile = destDir.resolve("release-notes.md").toFile
+  val artifactsFilter  = artifactExtensions.value.foldLeft[FileFilter](NothingFilter)((r, x) => r || GlobFilter(s"*.$x"))
+
+  val tags = runner("tag", "--sort", "version:refname")(cwd, sbt.Logger.Null).replaceAll("\r", "").split("\n")
+  val changesContent = if (tags.length <= 1) {
+    log.warn(s"Expected at least two tags, but there is ${tags.length}")
+    ""
+  } else {
+    val lastTag = tags.last
+    val prevTag = tags(tags.length - 2)
+    log.info(s"Looking for commits between $prevTag and $lastTag")
+
+    runner("log", "--pretty=%an%n%B%n", s"$prevTag..$lastTag")(cwd, sbt.Logger.Null).replaceAll("\r", "").replaceAll("\n{3,}", "\n").trim
+  }
+
+  val artifacts = IO.listFiles(destDir.toFile, artifactsFilter)
+  log.info(s"Found artifacts: ${artifacts.map(_.getName).mkString(", ")}")
+
+  val hashes = artifacts.map { file =>
+    val xs = Using.fileInputStream(file)(mk("SHA-256", _))
+    file.toPath.getFileName.toString -> Hex.encodeHexString(xs)
+  }
+
+  val sortedHashes = hashes.sortBy { case (fileName, _) => (fileName.length, fileName) }
+  val hashesContent =
+    s"""## SHA256 Checksums
+       |
+       |```
+       |${sortedHashes.map { case (fileName, hash) => s"$hash $fileName" }.mkString("\n")}
+       |```
+       |""".stripMargin
+
+  val content = s"""$changesContent
+                   |
+                   |$hashesContent""".stripMargin
+
+  log.info(s"Write release notes to $releaseNotesFile")
+  IO.write(releaseNotesFile, content, StandardCharsets.UTF_8)
+}
 
 lazy val checkPRRaw = taskKey[Unit]("Build a project and run unit tests")
 checkPRRaw := {
@@ -161,7 +220,17 @@ def checkPR: Command = Command.command("checkPR") { state =>
   state
 }
 
-commands += checkPR
+def packageAllForNetworks: Command = Command.command("packageAllForNetworks") { state =>
+  val testNetState = Project
+    .extract(state)
+    .appendWithoutSession(Seq(Global / network := Testnet), state)
+
+  Project.extract(state).runTask(root / packageAll, state)
+  Project.extract(testNetState).runTask(root / packageAll, testNetState)
+  state
+}
+
+commands ++= Seq(checkPR, packageAllForNetworks)
 
 // IDE settings
 ideExcludedDirectories := Seq((root / baseDirectory).value / "_local")
